Implement Sign-in api

+++ withaeng-api/src/main/kotlin/com/travel/withaeng/applicationservice/auth/AuthApplicationService.kt
@@ -1,12 +1,17 @@
 package com.travel.withaeng.applicationservice.auth
 
+import com.travel.withaeng.applicationservice.auth.dto.SignInServiceRequest
 import com.travel.withaeng.applicationservice.auth.dto.SignUpServiceRequest
 import com.travel.withaeng.applicationservice.auth.dto.UserResponse
+import com.travel.withaeng.common.exception.WithaengException
+import com.travel.withaeng.common.exception.WithaengExceptionType
 import com.travel.withaeng.domain.user.CreateUserDto
+import com.travel.withaeng.domain.user.UserRole
 import com.travel.withaeng.domain.user.UserService
 import com.travel.withaeng.domain.validateemail.ValidatingEmailService
 import com.travel.withaeng.security.authentication.UserInfo
 import com.travel.withaeng.security.jwt.JwtAgent
+import org.springframework.security.crypto.password.PasswordEncoder
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import java.util.*
@@ -16,14 +21,19 @@ import java.util.*
 class AuthApplicationService(
     private val userService: UserService,
     private val validatingEmailService: ValidatingEmailService,
-    private val jwtAgent: JwtAgent
+    private val jwtAgent: JwtAgent,
+    private val passwordEncoder: PasswordEncoder
 ) {
 
     @Transactional
     fun signUp(request: SignUpServiceRequest): UserResponse {
         val userEmail = request.email
         val userDto = userService.findByEmailOrNull(request.email)
+        val isValidUser = userDto?.roles?.any { it == UserRole.USER }
         if (userDto != null) {
+            if (isValidUser == true) {
+                throw WithaengException.of(WithaengExceptionType.ALREADY_EXIST, "이미 가입된 이메일입니다.")
+            }
             userService.deleteByEmail(userEmail)
         }
         val newUserDto = userService.createUser(request.toCreateUserDto())
@@ -31,9 +41,23 @@ class AuthApplicationService(
         return UserResponse(newUserDto.id, newUserDto.email, jwtAgent.provide(UserInfo.from(newUserDto)))
     }
 
+    @Transactional
+    fun signIn(request: SignInServiceRequest): UserResponse {
+        val userDto = userService.findByEmailOrNull(request.email)
+            ?: throw WithaengException.of(WithaengExceptionType.NOT_EXIST, "이메일에 해당하는 유저를 찾을 수 없습니다.")
+        checkValidUserPassword(request.password, userDto.password)
+        return UserResponse(userDto.id, userDto.email, jwtAgent.provide(UserInfo.from(userDto)))
+    }
+
+    private fun checkValidUserPassword(source: String, encryptedPassword: String) {
+        if (!passwordEncoder.matches(source, encryptedPassword)) {
+            throw WithaengException.of(WithaengExceptionType.AUTHENTICATION_FAILURE)
+        }
+    }
+
     private fun SignUpServiceRequest.toCreateUserDto(): CreateUserDto = CreateUserDto(
         email = email,
-        password = password,
+        password = passwordEncoder.encode(password),
         birth = birth,
         isMale = isMale
     )
